requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.20.3=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
argon2-cffi=20.1.0=py37h27cfd23_1
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=23.1.0=pypi_0
azure-ai-formrecognizer=3.2.1=pypi_0
azure-common=1.1.28=pypi_0
azure-core=1.28.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=4.1.0=pyhd3eb1b0_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cycler=0.11.0=pypi_0
dacite=1.8.1=pypi_0
datasets=2.13.1=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.6=pypi_0
donut-python=1.0.9=pypi_0
entrypoints=0.4=py37h06a4308_0
evaluate=0.4.0=pypi_0
filelock=3.12.2=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
google-auth=2.21.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.56.0=pypi_0
htmlmin=0.1.12=pypi_0
huggingface-hub=0.15.1=pypi_0
idna=3.4=pypi_0
imagehash=4.3.1=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
isodate=0.6.1=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=py37h06a4308_0
joblib=1.3.0=pypi_0
jsonschema=4.17.3=py37h06a4308_0
jupyter-core=4.12.0=pypi_0
jupyter_client=7.4.9=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.8.0=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=0.8.4=py37h14c3975_1001
msrest=0.7.1=pypi_0
multidict=6.0.4=pypi_0
multimethod=1.9.1=pypi_0
multiprocess=0.70.14=pypi_0
munch=3.0.0=pypi_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.1=py37h06a4308_0
nbformat=5.7.0=py37h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py37h06a4308_0
networkx=2.6.3=pypi_0
nltk=3.8.1=pypi_0
notebook=6.4.12=py37h06a4308_0
numpy=1.21.6=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.8=pypi_0
openssl=1.1.1u=h7f8727e_0
packaging=23.1=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.3=pypi_0
peft=0.3.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
phik=0.12.3=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=py37h06a4308_0
plotly=5.15.0=pypi_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=12.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.1.0=pypi_0
pyrsistent=0.18.0=py37heee7806_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py37h06a4308_0
pytorch-lightning=1.9.5=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.6.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rouge-score=0.1.2=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.32=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
safetensors=0.3.1=pypi_0
scipy=1.7.3=pypi_0
sconf=0.2.5=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentencepiece=0.1.99=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
statsmodels=0.13.5=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.17.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
timm=0.6.13=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch=1.13.1=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py37h5eee18b_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.25.1=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.7.0=pypi_0
urllib3=1.26.16=pypi_0
visions=0.7.5=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=py37_1
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=3.5.2=py37h06a4308_0
wordcloud=1.9.2=pypi_0
xxhash=3.2.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
ydata-profiling=4.3.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zss=1.2.0=pypi_0
